Wired up a KeyBasedCrypterCache for EAR (#57)
* Wired up a KeyBasedCrypterCache for EAR

This commmit adds a cache object that is updated during segment load (or reload).
The cache object is made available to the query engine via QueryContext, so that
the transform functions can use the object.

The idea is that when segments are loaded, we go through all the values of the
Key column (specified in tableConfig -- exact syntax TBD), and fetch the dcryption
key for each value from KMS. We then construct the decrypter object and cache it.

When a transform function is invoked (e.g. to decrypt a column), the transform function
has access to the cache that has a pre-constructed crypter object. The crypter object
can be used to decrypt values during query processing.

The cache itself is yet to be implemented. It is expected that the object handles
modification while queries are being executed (e.g. segment can be reloaded while
query is being executed in the same segment).

The segment load (or reload) object needs to wait until the all the objects needed
are cached. The exact mechanism for this is TBD.

* Fix style check

* More Style checks fixed

* Spotless apply

* More style fixes

* Removed test code