Added transform/rewrite functionality (#69)
* Added transform/rewrite functionality

    The query rewriter for EAR needs to know the names and types of key column
    and encrypted columns, so that it can rewrite to the appropriate method
    implemented (or throw exception that we cannot decrypt the column). Added a rewriter
    and wired up the rewriter to get notified when tables get added to or removed from
    the broker.

    Added a transform function for columns that encrypt double or long values. These
    should be good for phase-1 deployment (unless we decide to use String as encrypted
    format -- in which case we will need to change the arguments to be of String type
    instead of byte array).

    TODO adding tests for decrypt transform functions, once we have the row-level crypter
    ready

* Added a crypter interface and default crypter

For now, the coding in the decryption path allows the encrypted column
to be a base64 encoded string. This helps us get to phase-1 easier, where
the customer already has encrypted values stored as String.

* Fixed bugs in rewriter initialization

* Removed multiple hashes in DecryptRewriter

Addressing the review comment to introduce an EarInfo object instead
of having multiple hash tables in the rewriter. Re-factored tests and
added some more tests.

* Added TableQueryRewriter

Added a new interface that query rewriters can potentially implement to
get notified of table registration and de-registration.

* Removing commented code